# docs-back/swagger.yaml
basePath: /api/v1
definitions:
...
  create_monitoring_config.CreateMonitoringConfigRequest:
    properties:
      block_ads_cookies:
        type: boolean
      check_frequency:
        description: '"5m", "1h", "1d"'
        type: string
      page_id:
        type: string
      schedule_type:
        description: '"continuous", "scheduled"'
        type: string
      timezone:
        description: '"UTC", "America/New_York", etc'
        type: string
    type: object
  create_monitoring_config.CreateMonitoringConfigResponse:
    properties:
      block_ads_cookies:
        type: boolean
      check_frequency:
        type: string
      created_at:
        type: string
      id:
        type: string
      page_id:
        type: string
      schedule_type:
        type: string
      timezone:
        type: string
    type: object
  create_notification_preference.CreateNotificationPreferenceRequest:
    properties:
      change_types:
        items:
          type: string
        type: array
      email_enabled:
        type: boolean
      page_id:
        type: string
      user_id:
        type: string
      workspace_id:
        type: string
    type: object
  create_notification_preference.CreateNotificationPreferenceResponse:
    properties:
      change_types:
        items:
          type: string
        type: array
      created_at:
        type: string
      email_enabled:
        type: boolean
      id:
        type: string
      page_id:
        type: string
      user_id:
        type: string
      workspace_id:
        type: string
    type: object
...
  createcheck.CreateCheckRequest:
    properties:
      change_detected:
        type: boolean
      change_type:
        type: string
      duration_ms:
        type: integer
      error_message:
        type: string
      html_snapshot_url:
        type: string
      page_id:
        type: string
      screenshot_url:
        type: string
      status:
        type: string
    required:
    - page_id
    - status
    type: object
  createcheck.CreateCheckResponse:
    properties:
      change_detected:
        type: boolean
      change_type:
        type: string
      checked_at:
        type: string
      id:
        type: string
      page_id:
        type: string
      status:
        type: string
    type: object
...
  updatemonitoringconfig.UpdateMonitoringConfigRequest:
    properties:
      block_ads_cookies:
        type: boolean
      check_frequency:
        type: string
      schedule_type:
        type: string
      timezone:
        type: string
    type: object
  updatemonitoringconfig.UpdateMonitoringConfigResponse:
    properties:
      block_ads_cookies:
        type: boolean
      check_frequency:
        type: string
      id:
        type: string
      page_id:
        type: string
      schedule_type:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:9090
info:
  contact:
    email: support@pulzifi.com
    name: API Support
  description: Pulzifi backend monolith API
  license:
    name: MIT
  title: Pulzifi API
  version: "1.0"
paths:
...
  /monitoring/checks:
    get:
      description: List all monitoring checks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Monitoring Checks
      tags:
      - monitoring
    post:
      consumes:
      - application/json
      description: Create a new monitoring check
      parameters:
      - description: Create Check Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/createcheck.CreateCheckRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/createcheck.CreateCheckResponse'
      security:
      - BearerAuth: []
      summary: Create Monitoring Check
      tags:
      - monitoring
  /monitoring/checks/{id}:
    get:
      description: Get a monitoring check by ID
      parameters:
      - description: Check ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Monitoring Check
      tags:
      - monitoring
  /monitoring/configs:
    post:
      consumes:
      - application/json
      description: Create a new monitoring config
      parameters:
      - description: Create Config Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/create_monitoring_config.CreateMonitoringConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/create_monitoring_config.CreateMonitoringConfigResponse'
      security:
      - BearerAuth: []
      summary: Create Monitoring Config
      tags:
      - monitoring
  /monitoring/configs/{pageId}:
    get:
      description: Get a monitoring config by page ID
      parameters:
      - description: Page ID
        in: path
        name: pageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Monitoring Config
      tags:
      - monitoring
    put:
      consumes:
      - application/json
      description: Update an existing monitoring config or create a new one if it
        doesn't exist
      parameters:
      - description: Page ID
        in: path
        name: pageId
        required: true
        type: string
      - description: Update Config Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/updatemonitoringconfig.UpdateMonitoringConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/updatemonitoringconfig.UpdateMonitoringConfigResponse'
      security:
      - BearerAuth: []
      summary: Update or Create Monitoring Config (Upsert)
      tags:
      - monitoring
  /monitoring/notification-preferences:
    post:
      consumes:
      - application/json
      description: Create a new notification preference
      parameters:
      - description: Create Preference Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/create_notification_preference.CreateNotificationPreferenceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/create_notification_preference.CreateNotificationPreferenceResponse'
      security:
      - BearerAuth: []
      summary: Create Notification Preference
      tags:
      - monitoring
  /monitoring/notification-preferences/{id}:
    get:
      description: Get a notification preference by ID
      parameters:
      - description: Preference ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Notification Preference
      tags:
      - monitoring
...
schemes:
- http
- https
swagger: "2.0"